Make top-level section work with null PerfCake model
Create PerfClipse model for top level section also if PerfCake model for
these section is null. This will allow to display these sections in
editor. But setters needed to be fixed in a way that if something is
added then at first PerfCake model must be created (if it is null) and
then perform required action.